+++ tests/test_callbacks.py
@@ -7,18 +7,17 @@ import pytest
 from pytorch_lightning.callbacks import ModelCheckpoint
 from deepforest import get_data
 
-def test_log_images(m, tmpdir):
-    im_callback = callbacks.images_callback(csv_file=m.config["validation"]["csv_file"], root_dir=m.config["validation"]["root_dir"], savedir=tmpdir)
+@pytest.mark.parametrize("every_n_epochs", [1, 2, 3])
+def test_log_images(m, every_n_epochs, tmpdir):
+    im_callback = callbacks.images_callback(savedir=tmpdir, every_n_epochs=every_n_epochs)
     m.create_trainer(callbacks=[im_callback])
     m.trainer.fit(m)
     saved_images = glob.glob("{}/*.png".format(tmpdir))
     assert len(saved_images) == 1
     
-def test_log_images_multiclass(two_class_m, tmpdir):
-    im_callback = callbacks.images_callback(
-        csv_file=two_class_m.config["validation"]["csv_file"],
-        root_dir=two_class_m.config["validation"]["root_dir"],
-        savedir=tmpdir)
+@pytest.mark.parametrize("every_n_epochs", [1, 2, 3])
+def test_log_images_multiclass(two_class_m, every_n_epochs, tmpdir):
+    im_callback = callbacks.images_callback(savedir=tmpdir, every_n_epochs=every_n_epochs)
     two_class_m.create_trainer(callbacks=[im_callback])
     two_class_m.trainer.fit(two_class_m)
     saved_images = glob.glob("{}/*.png".format(tmpdir))